% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{my_multi_lm}
\alias{my_multi_lm}
\title{Descriptions about three main functions:
\enumerate{
\item my_multi_lm
Used for fitting a multiple linear regression model and compute related statistics
Usage Example: result <- my_multi_lm(y ~ x1 + x2 + x3, data = your_data_frame)
}}
\usage{
my_multi_lm(formula, data)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}} specifying the model.}

\item{data}{A data frame containing the data to be modeled.}
}
\value{
A list containing model outputs such as coefficients, fitted values,
residuals, sigma hat, hat values, mean squared error, R-squared, and adjusted R-squared,
along with the model matrix.
}
\description{
\enumerate{
\item multi_lm_diag
Used for conducting diagnostic analysis of the model, including calculating residuals, and identifying outliers
Usage Example: diagnostics <- multi_lm_diag(model_results = result)
}
}
\details{
Note:When implementing these functions, ensure that the model_results passed from my_multi_lm to the other functions contains all necessary data (e.g., model matrix, residuals, coefficients) to allow for full functionality in diagnostics and plotting.

This function fits a multiple linear regression model to the provided data and computes various statistical measures including coefficients, residuals, and R-squared values.
}
\examples{
set.seed(123)
data <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = rnorm(100))
result <- my_multi_lm(y ~ x1 + x2, data = data)
print(result)
}
